+++ package.json
@@ -21,8 +21,9 @@
     "@reach/rect": "^0.10.5",
     "@sindresorhus/slugify": "^1.1.0",
     "@svgr/webpack": "^5.4.0",
+    "@tailwindcss/jit": "^0.1.18",
     "@tailwindcss/typography": "^0.0.0-caa70b6",
-    "autoprefixer": "^10.0.2",
+    "autoprefixer": "^10.2.5",
     "babel-plugin-preval": "^5.0.0",
     "brotli-size": "^4.0.0",
     "clean-css": "^4.2.3",
@@ -53,7 +54,7 @@
     "seedrandom": "^3.0.5",
     "simple-functional-loader": "^1.2.1",
     "stringify-object": "^3.3.0",
-    "tailwindcss": "^2.0.3",
+    "tailwindcss": "^2.0.4",
     "tinytime": "^0.2.6",
     "unist-util-visit": "^2.0.3"
   },
