+++ src/utils/stats.json
@@ -1,0 +1,1 @@
+{"1-screen":{"original":"1141.7kB","minified":"935.6kB","gzipped":"96.8kB","brotlified":"50.7kB"},"2-screens":{"original":"1742.7kB","minified":"1414.7kB","gzipped":"145.0kB","brotlified":"55.3kB"},"25-colors":{"original":"2098.6kB","minified":"1639.0kB","gzipped":"189.9kB","brotlified":"47.1kB"},"3-screens":{"original":"2343.9kB","minified":"1894.0kB","gzipped":"192.8kB","brotlified":"60.9kB"},"35-colors-3-screens":{"original":"1543.1kB","minified":"1218.8kB","gzipped":"137.7kB","brotlified":"44.3kB"},"50-colors":{"original":"2726.9kB","minified":"2167.5kB","gzipped":"231.9kB","brotlified":"57.2kB"},"default":{"original":"3566.2kB","minified":"2872.2kB","gzipped":"289.2kB","brotlified":"71.3kB"}}
